adding data validation

+++ Controllers/CommentController.cs
@@ -24,6 +24,8 @@ namespace cs_webapi.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             var comments = await _commentRepository.GetAllCommentsAsync();
             var commentDto = comments.Select(c => c.ToCommentDto());
 
@@ -35,9 +37,11 @@ namespace cs_webapi.Controllers
         }
 
         [HttpGet]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             var comment = await _commentRepository.GetCommentById(id);
 
             if(comment == null) {
@@ -47,9 +51,11 @@ namespace cs_webapi.Controllers
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentRequestDto commentRequestDto)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             if(!await _stockRepository.StockExists(stockId)) {
                 return BadRequest("Stock not found");
             }
@@ -61,9 +67,11 @@ namespace cs_webapi.Controllers
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateCommentRequestDto)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             var commentModel = await _commentRepository.UpdateCommentAsync(id, updateCommentRequestDto.ToCommentFromUpdateDTO());
 
             if(commentModel == null) {
@@ -74,9 +82,11 @@ namespace cs_webapi.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+            
             var commentModel = await _commentRepository.DeleteCommentAsync(id);
 
             if(commentModel == null) {
