+++ Controllers/StockController.cs
@@ -26,6 +26,8 @@ namespace cs_webapi.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             var stocks = await _stockRepository.GetAllStockAsync();
             
             var stockDto = stocks.Select(s => s.ToStockDto());
@@ -37,9 +39,11 @@ namespace cs_webapi.Controllers
             return Ok(stocks);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             var stock = await _stockRepository.GetStockById(id);
 
             if(stock == null) {
@@ -52,6 +56,8 @@ namespace cs_webapi.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockRequestDto)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             var stockModel = stockRequestDto.ToStockFromCreateDTO();
             await _stockRepository.CreateStockAsync(stockModel);
 
@@ -59,9 +65,11 @@ namespace cs_webapi.Controllers
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateStockRequestDto)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             var stockModel = await _stockRepository.UpdateStockAsync(id, updateStockRequestDto);
 
             if(stockModel == null) {
@@ -72,9 +80,11 @@ namespace cs_webapi.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id}:int")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+            
             var stockModel = await _stockRepository.DeleteStockAsync(id);
 
             if(stockModel == null) {
